generation/requirements.txt
python-dotenv>=1.0.0
mutagen>=1.47.0
pydub>=0.25.0
audioop-lts>=0.2.0; python_version >= "3.13"
openai>=1.0.0
litellm>=1.0.0
pyyaml>=6.0
requests>=2.28.0
elevenlabs>=1.0.0
faster-whisper>=1.0.0
google-api-python-client>=2.0.0
google-api-python-client-stubs>=1.31.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
